# data/fossology-LicenseDb/cmd/laas/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service to host license information for other services to query over REST API.",
        "title": "laas (License as a Service) API",
        "contact": {
            "name": "FOSSology",
            "url": "https://fossology.org",
            "email": "fossology@fossology.org"
        },
        "license": {
            "name": "GPL-2.0-only",
            "url": "https://github.com/fossology/LicenseDb/blob/main/LICENSE"
        },
        "version": "0.0.9"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/obligations/categories": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "{}": []
                    }
                ],
                "description": "Get all active obligation categories from the service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Obligations"
                ],
                "summary": "Get all active obligation categories",
                "operationId": "GetAllObligationCategories",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Active obligation categories only",
                        "name": "active",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ObligationCategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Unable to fetch obligation categories",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    },
                    "500": {
                        "description": "Unable to fetch obligation categories",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create an obligation category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Obligations"
                ],
                "summary": "Create an obligation category",
                "operationId": "CreateObligationCategory",
                "parameters": [
                    {
                        "description": "Obligation category to create",
                        "name": "obligation_category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ObligationCategory"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ObligationCategoryResponse"
                        }
                    },
                    "400": {
                        "description": "invalid json body",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    },
                    "409": {
                        "description": "obligation category already exists",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    },
                    "500": {
                        "description": "something went wrong while creating new obligation category",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            }
        },
        "/obligations/categories/{category}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deactivate an obligation category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Obligations"
                ],
                "summary": "Deactivate obligation category",
                "operationId": "DeleteObligationCategory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Obligation Category",
                        "name": "category",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "cannot delete obligation category 'DISTRIBUTION' as it's still referenced by some obligations",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    },
                    "404": {
                        "description": "obligation category 'DISTRIBUTION' not found",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    },
                    "500": {
                        "description": "something went wrong while deleting obligation category",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.LicenseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request body"
                },
                "message": {
                    "type": "string",
                    "example": "invalid request body"
                },
                "path": {
                    "type": "string",
                    "example": "/api/v1/licenses"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-12-01T10:00:51+05:30"
                }
            }
        },
...
        "models.ObligationCategory": {
            "type": "object",
            "required": [
                "category"
            ],
            "properties": {
                "category": {
                    "type": "string",
                    "example": "GENERAL"
                }
            }
        },
        "models.ObligationCategoryResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ObligationCategory"
                    }
                },
                "paginationmeta": {
                    "$ref": "#/definitions/models.PaginationMeta"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "models.PaginationMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "next": {
                    "type": "string",
                    "example": "/api/v1/licenses?limit=10\u0026page=11"
                },
                "page": {
                    "type": "integer",
                    "example": 10
                },
                "previous": {
                    "type": "string",
                    "example": "/api/v1/licenses?limit=10\u0026page=9"
                },
                "resource_count": {
                    "type": "integer",
                    "example": 200
                },
                "total_pages": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Token from /login endpoint. Enter the token with the `Bearer ` prefix, e.g. \\\"Bearer eyJhbGciOiJ.....\\\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}